requirements.txt
requests
beautifulsoup4
SQLAlchemy
pyTelegramBotAPI
python-dotenv
psycopg2-binary
